исправил баги и доработал логику

+++ Server/LifeGrind.CORE/Services/SkillService.cs
@@ -1,17 +1,25 @@
+using LifeGrind.CORE.Exceptions;
+
 public class SkillService : ISkillService
 {
     private readonly ISkillRepository skillRepository;
+
     public SkillService(ISkillRepository _SkillRepository)
     {
         skillRepository = _SkillRepository;
     }
-    public async Task<List<SkillEntity>> GetSkills(){
+
+    public async Task<List<SkillEntity>> GetSkills()
+    {
         return await skillRepository.GetSkills();
     }
-    public async Task<List<SkillEntity>> GetSkillsByUserId(Guid userId){
+
+    public async Task<List<SkillEntity>> GetSkillsByUserId(Guid userId)
+    {
         return await skillRepository.GetSkillsByUserId(userId);
     }
-    public async Task Add(Guid userId,CreateSkillRequest request)
+
+    public async Task Add(Guid userId, CreateSkillRequest request)
     {
         var skill = new SkillEntity()
         {
@@ -21,17 +29,39 @@ public class SkillService : ISkillService
             Experience = request.Experience,
             UserId = userId
         };
+
         await skillRepository.Add(skill);
     }
 
-    
-    public async Task Update(Guid skillId, UpdateSkillRequest request)
+    public async Task Update(Guid userId, Guid skillId, UpdateSkillRequest request)
     {
-        await skillRepository.Update(skillId,request.Name,request.Description,request.Experience);
-    }
-    public async Task Delete(Guid skillId){
-        await skillRepository.Delete(skillId);
+        var skill = await skillRepository.GetSkill(skillId);
+        if (skill == null)
+        {
+            throw new SkillNotExistException();
+        }
+
+        if (skill.UserId != userId)
+        {
+            throw new EntityNotFoundException("Attempt to update another user's skill.");
+        }
+
+        await skillRepository.Update(skillId, request.Name, request.Description, request.Experience);
     }
 
-    // Все проверки тут будут с FV, так что тут по идее мне и нечего писать
+    public async Task Delete(Guid userId, Guid skillId)
+    {
+        var skill = await skillRepository.GetSkill(skillId);
+        if (skill == null)
+        {
+            throw new SkillNotExistException();
+        }
+
+        if (skill.UserId != userId)
+        {
+            throw new EntityNotFoundException("Attempt to delete another user's skill.");
+        }
+
+        await skillRepository.Delete(skillId);
+    }
 }
